refactor(mdast): split writer.rs into a writer/ module directory
writer.rs had grown to 1547 lines mixing the public API surface, the
Writer struct's core dispatch, and its block-level and inline-level
node-writing methods. Split into writer/{mod,block,inline,tests}.rs,
splitting the Writer impl block across sibling files (fields stay
private in mod.rs, accessible to child modules per Rust's privacy
rules). No behavior change and no public API change -- every item
lib.rs re-exports from writer:: still resolves at the same path.
